+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unpublished
+### Fixed
+- Fixed `save_spectra()` truncating an existing file when called with an empty spectrum list and `append=True`.
 
 ## [0.31.0] - 2025-10-06
 ### Added